refactors preview search to stand alone
